Restrukturerer funksjoner for server og route-handlers
