+++ server/server.ts
@@ -1,194 +1,52 @@
-import express, { Request } from 'express';
+import express from 'express';
 import cors from 'cors';
-import { DataKeys, buildDataStructure } from '@/utils';
-import { Index } from '@/views/index';
-import { Footer } from '@/views/footer';
-import { HeaderMenuLinks } from '@/views/header-menu-links';
-import { Header } from '@/views/header';
+
 import { decoratorParams } from '@/server/middlewares';
-import { mockAuthHandler } from '@/server/mock/authMock';
+import { isAliveHandler, isReadyHandler } from './handlers/livenessHandlers';
 import {
+  driftsmeldingerHandler,
+  searchHandler,
+} from './handlers/serviceHandlers';
+import { dataHandlers, inspectData } from './handlers/dataHandlers';
+import {
+  mockAuthHandler,
   mockSessionHandler,
   refreshMockSessionHandler,
-} from '@/server/mock/sessionMock';
-import { DecoratorEnv } from '@/views/decorator-env';
-import { DecoratorLens } from '@/decorator-lens';
-import { isAliveHandler, isReadyHandler } from './common';
-
-import { driftsmeldingerHandler } from './api/driftsmeldinger';
+} from '@/server/handlers/mockHandlers';
+import {
+  footerHandler,
+  headerHandler,
+  indexHandler,
+} from './handlers/domHandlers';
 
 const isProd = process.env.NODE_ENV === 'production';
 const port = process.env.PORT;
-const host = process.env.HOST ?? `http://localhost:${port}`;
-
-const entryPointPath = 'client/main.ts';
-
-const script = (src: string) => `<script type="module" src="${src}"></script>`;
-
-const getResources = async () => {
-  const resources = (
-    (await import('../dist/manifest.json', { assert: { type: 'json' } }))
-      .default as {
-      [entryPointPath]: { file: string; css: string[] };
-    }
-  )[entryPointPath];
-
-  if (isProd) {
-    return {
-      scripts: script(`${host}/${resources.file}`),
-      styles: [
-        ...resources.css.map(
-          (href: string) =>
-            `<link type="text/css" rel="stylesheet" href="${host}/${href}"></link>`,
-        ),
-      ].join(''),
-    };
-  }
-
-  return {
-    styles: '',
-    scripts: [
-      'http://localhost:5173/@vite/client',
-      `http://localhost:5173/${entryPointPath}`,
-    ]
-      .map(script)
-      .join(''),
-  };
-};
-
-const resources = await getResources();
 const app = express();
 
+// Setup middleware
 app.use(cors());
 app.use(express.static(isProd ? 'dist' : 'public'));
 app.use(decoratorParams);
 
+// Liveness and mock handlers
 app.use('/api/isReady', isReadyHandler);
 app.use('/api/isAlive', isAliveHandler);
 app.use('/api/auth', mockAuthHandler);
 app.get('/api/oauth2/session', mockSessionHandler);
 app.get('/api/oauth2/session/refresh', refreshMockSessionHandler);
 
-type SearchHit = {
-  displayName: string;
-  highlight: string;
-  href: string;
-};
-
-type SearchResponse = {
-  c: number;
-  isMore: boolean;
-  s: number;
-  word: string;
-  total: number;
-  hits: SearchHit[];
-};
-
-app.use('/dekoratoren/api/sok', async (req: Request<{ ord: string }>, res) => {
-  const results = (await (
-    await fetch(`https://www.nav.no/dekoratoren/api/sok?ord=${req.query.ord}`)
-  ).json()) as SearchResponse;
-
-  res.json({
-    hits: results.hits.slice(0, 5),
-    total: results.total,
-  });
-});
-
+// Service handlers
+app.use('/dekoratoren/api/sok', searchHandler);
 app.use('/dekoratoren/api/driftsmeldinger', driftsmeldingerHandler);
 
-app.use('/footer', async (req, res) => {
-  const params = req.decoratorParams;
-  // Maybe make into middleware
-  const data = await buildDataStructure(params);
+// Data handlers
+app.use('/data/inspect-data', inspectData);
+app.get('/data/:key', dataHandlers);
 
-  return res.status(200).send(
-    Footer({
-      simple: req.decoratorParams.simple,
-      personvern: data.personvern,
-      footerLinks: data.footerLinks,
-      feedback: req.decoratorParams.feedback,
-      texts: data.texts,
-    }),
-  );
-});
-
-app.use('/inspect-data', async (req, res) => {
-  const data = await buildDataStructure(req.decoratorParams);
-  const raw = await fetch('https://www.nav.no/dekoratoren/api/meny');
-  res.json({
-    data,
-    raw: await raw.json(),
-  });
-});
-
-app.use('/header', async (req, res) => {
-  const params = req.decoratorParams;
-  const data = await buildDataStructure(params);
-  return res.status(200).send(
-    HeaderMenuLinks({
-      headerMenuLinks: data.headerMenuLinks,
-    }),
-  );
-});
-
-app.get('/data/:key', async (req, res) => {
-  const { params } = req;
-  const dataKey = params.key as DataKeys;
-
-  if (!dataKey) {
-    return res.status(400).send('Missing key');
-  }
-
-  const data = await buildDataStructure(req.decoratorParams);
-  const subset = data[dataKey];
-
-  if (!subset) {
-    res.status(404).send('Data not found with key:' + dataKey);
-  }
-
-  res.send(subset);
-});
-
-app.use('/', async (req, res) => {
-  const data = await buildDataStructure(req.decoratorParams);
-  const fullUrl = req.protocol + '://' + req.get('host');
-
-  res.status(200).send(
-    Index({
-      scripts: resources.scripts,
-      links: resources.styles,
-      language: req.decoratorParams.language,
-      header: Header({
-        texts: data.texts,
-        mainMenu: data.mainMenu,
-        headerMenuLinks: data.headerMenuLinks,
-        innlogget: false,
-        isNorwegian: true,
-        breadcrumbs: req.decoratorParams.breadcrumbs,
-        utilsBackground: req.decoratorParams.utilsBackground,
-        availableLanguages: req.decoratorParams.availableLanguages,
-        myPageMenu: data.myPageMenu,
-      }),
-      footer: Footer({
-        texts: data.texts,
-        personvern: data.personvern,
-        footerLinks: data.footerLinks,
-        simple: req.decoratorParams.simple,
-        feedback: req.decoratorParams.feedback,
-      }),
-      env: DecoratorEnv({
-        origin: fullUrl,
-        env: req.decoratorParams,
-      }),
-      lens: DecoratorLens({
-        origin: fullUrl,
-        env: req.decoratorParams,
-        query: req.query,
-      }),
-    }),
-  );
-});
+// DOM handlers
+app.use('/footer', footerHandler);
+app.use('/header', headerHandler);
+app.use('/', indexHandler);
 
 app.listen(port, function () {
   console.log(`Listening on http://localhost:${port}`);
